// src/modules/product-variant/constants/product-variant.enums.ts
export enum ProductVariantErrorKeyEnum {
  PRODUCT_ID_IS_REQUIRED = 'Product ID is required',
  FAILED_TO_FETCH_PRODUCT = 'Failed to fetch product',
  FAILED_TO_CREATE_PRODUCT_VARIANT = 'Failed to create product variant',
  FAILED_VARIANT_RECORD = 'Failed variant record',
  FAILED_TO_INSERT_PRODUCT_IMAGES = 'Failed to insert product images',
  FAILED_TO_FETCH_PRODUCT_VARIANTS_BY_PRODUCT_ID = 'Failed to fetch product variants by product ID',
  PRODUCT_VARIANT_NOT_FOUND = 'Product variant not found',
  PRODUCT_VARIANT_IS_INACTIVE = 'Product variant is inactive',
  FAILED_TO_FETCH_PRODUCT_VARIANT_DETAILS = 'Failed to fetch product variant details',
  FAILED_TO_FETCH_PRODUCT_VARIANTS = 'Failed to fetch product variants',
  FAILED_TO_FETCH_PRODUCT_VARIANT = 'Failed to fetch product variant',
  COMPANY_DOMAIN_IS_REQUIRED = 'Company domain is required',
  FAILED_TO_UPDATE_PRODUCT_VARIANT = 'Failed to update product variant',
  FAILED_TO_FETCH_STOCK_MANAGER_DATA = 'Failed to fetch stock manager data',
  PRODUCT_STATUS_IS_REQUIRED = 'Product status is required',
  FAILED_TO_UPDATE_PRODUCT_VARIANT_STATUS = 'Failed to update product variant status',
  FAILED_TO_UPDATE_PRODUCT_STATUS = 'Failed to update product status',
  ID_REQUIRED = 'id required',
  FAILED_TO_DELETE_PRODUCT_VARIANT = 'Failed to delete product variant',
}